Improve user shift management
